+++ src/Pwa/AssetManifest.php
@@ -860,9 +860,9 @@ class AssetManifest
             (string) json_encode($this->config->get('pwax.blade', [])),
             (string) json_encode($this->config->get('pwax.styles', [])),
             (string) json_encode($this->config->get('pwax.scripts', [])),
-            (string) json_encode($this->config->get('pwax.plugins', [])),
-            (string) json_encode($this->config->get('pwax.directives', [])),
-            (string) json_encode($this->config->get('pwax.client_middleware', $this->config->get('pwax.middleware_js', []))),
+            (string) json_encode($this->config->get('pwax.vue.plugins', [])),
+            (string) json_encode($this->config->get('pwax.vue.directives', [])),
+            (string) json_encode($this->config->get('pwax.vue.middleware', [])),
             $this->webManifest->hash(),
             (string) $this->hashFile(dirname(__DIR__, 2) . '/dist/pwax.js'),
         ];